backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
multidict==6.0.5
pdfkit==1.0.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
yarl==1.9.4
